reorganize files and components
