+++ client/src/components/Checkout/CheckoutSuccess.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const CheckoutSuccess = () => {
+  return <div>success</div>;
+};
+
+export default CheckoutSuccess;
