+++ client/src/pages/Checkout.js
@@ -4,7 +4,8 @@ import { loadStripe } from "@stripe/stripe-js";
 import { Elements } from "@stripe/react-stripe-js";
 import axios from "axios";
 
-import CheckoutForm from "../components/CheckoutForm";
+import CheckoutForm from "../components/Checkout/CheckoutForm";
+import CheckoutSuccess from "../components/Checkout/CheckoutSuccess";
 
 const stripePromise = loadStripe(
   "pk_test_51IAF0CHUsZNgCog2HKJ6N7blSBXlpTbMyPOsW4bXMVHlrWEPAbvhkcLtHxrcLdB7Git73G7i4eU2I4kovKAfBhvY00gbMjxWER"
@@ -15,6 +16,7 @@ const CheckoutPage = () => {
 
   const [amount, setAmount] = useState("");
   const [clientSecret, setClientSecret] = useState("");
+  const [status, setStatus] = useState(""); //pending, success, failure
 
   useEffect(() => {
     axios
@@ -36,24 +38,17 @@ const CheckoutPage = () => {
 
   return (
     <Elements stripe={stripePromise}>
-      <Paper className={classes.root} elevation={5} spacing={5}>
-        <Grid
-          className={classes.grid}
-          container
-          direction="column"
-          justify="space-around"
-          alignItems="center"
-        >
-          <Grid item>
-            <Typography variant="h5" className={classes.title}>
-              Checkout
-            </Typography>
-          </Grid>
-
-          <Grid item>
-            <CheckoutForm amount={amount} clientSecret={clientSecret} />
-          </Grid>
-        </Grid>
+      <Paper className={classes.root} elevation={5}>
+        {status === "success" ? (
+          <CheckoutSuccess />
+        ) : (
+          <CheckoutForm
+            amount={amount}
+            clientSecret={clientSecret}
+            status={status}
+            setStatus={setStatus}
+          />
+        )}
       </Paper>
     </Elements>
   );
@@ -61,15 +56,10 @@ const CheckoutPage = () => {
 
 const useStyles = makeStyles((theme) => ({
   root: {
-    height: theme.spacing(40),
+    height: theme.spacing(50),
     width: theme.spacing(60),
     margin: `${theme.spacing(10)}px auto`,
-  },
-  form: {
-    width: theme.spacing(40),
-  },
-  grid: {
-    height: "100%",
+    position: "relative",
   },
 }));
